Added a bottom navbar
This navbar contains 5 navitgation items:
- Search
- My events
- Pinpoint
- Overview
- Contacts

The 'overview' page wasn't discussed, but I added it because it might come in handy as some sort of landing page.

IMPORTANT: Since creating a branch for each navigationable screen would be quite the chaos for such a small task, I included these creations in this pull. The following issues are included in this branch:
- #11
- #17
- #23
- #30