Map format research + implementation
I researched a couple of methods for implementing the custom maps. For this, I found the 'flutter_map' library that seems perfect for this use case.

I created a new page ('map_tryout'), where a map is shown. I also added two custom maps (Rock Werchter & Pukkelpop) to show what an overlay looks like.

BUG ALERT: I noticed that when the location or map_tryout widget are displayed, clicking on the pinpoint menu item doesn't navigate back to the pinpoint landing. However, when first clicking on a different menu item and then on the pinpoint one, the expected behaviour is achieved.

I assume that the bug originates from the widget not being reloaded/resetted when the pinpoint menu item is clicked, since it's already being displayed. I will create a ticket for this bug fix.